+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 - Allow usage of large Survey Guide files (more than 2MB)
 
+## [1.7.3] - build 162 2019-05-07
+### Added
+- Support Interview Label
+
+## [1.7.3] - build 161 2019-04-30
+### Fixed
+- Allow remote Collect server connection in Android 8+
+
 ## [1.7.3] - build 160 2019-04-18
 ### Fixed
 - Record creation and modified dates in data export